Added alert confirmation window

+++ public/js/dashboard.js
@@ -25,24 +25,30 @@ const newFormHandler = async (event) => {
 
 // Handler for deleting a post from the dashboard
 const deleteButtonHandler = async (event) => {
-  // Ensure the event is being triggered by the correct button with the data-id
-  if (event.target.hasAttribute("data-id")) {
-    const id = event.target.getAttribute("data-id");
-    console.log(id);
 
-    // Use the API call for deleting a post
-    const response = await fetch(`/api/posts/${id}`, {
-      method: "DELETE",
-    });
+  var delConf = confirm("Are you sure you would like to delete this post?");
+  if (delConf == true) {
+    // Ensure the event is being triggered by the correct button with the data-id
+    if (event.target.hasAttribute("data-id")) {
+      const id = event.target.getAttribute("data-id");
 
-    if (response.ok) {
-      // Refresh page with updated list of posts
-      document.location.replace("/dashboard");
-    } else {
-      // If there's an error, say it
-      alert("Failed to delete post");
+      // Use the API call for deleting a post
+      const response = await fetch(`/api/posts/${id}`, {
+        method: "DELETE",
+      });
+
+      if (response.ok) {
+        // Refresh page with updated list of posts
+        document.location.replace("/dashboard");
+      } else {
+        // If there's an error, say it
+        alert("Failed to delete post");
+      }
     }
+  } else {
+    return false;
   }
+
 };
 
 document
